Audio Changes
Added a template theme to Boss Room 1 for testing of the audio.
Added a settings menu in Boss Room 1.
Pause Menu (if not mentioned before) was added with each button working as intended.
Added a MasterVolume Slider in Boss Room 1 along with an active mute button.
MasterVolume Slider has a save feature that saves the last audio volume when exited.
Moved some of the scripts into the script folder for better clarity on their location. Will likely add folders within the script folder to help divide the scripts into more identifiable sections.