Commits. (Read description for more info).
Added folders and moved the files I was personally using into them for better organization.
Created a Health Bar. Works and simply needs to be linked with the actual player's health to change and slide.
Created a coin counter. Needs to be worked on to count coins when the function is implemented.
Added a SFX when clicking the play button in the start menu.