Look at description.
Added dialogue to the Spawn Room. This includes a dialogue canvas and a dialogue script that prints the characters once at a time, a fast froward using a click, and pauses the game while the dialog is up.
I also added the PauseMenu (and it's other aspects) along with the OverworldAudio to the scene.
Additionally, I placed the game objects within the spawnroom scene into on empty object for organization and added more music files.