Look at description.
Added some SFX files.
Got rid of the options button in the main menu.
Adjusted the audio sound in both the main menu and the Overworld.
Added a fade out when quitting to main menu.
Created a new UI Document that holds the old UI in the case that I go back to fix it.