Upgrade cards and buying cards, Created a player deck
Got the card shop to add cards to the player deck that had been created in the spawn room. the upgrade cards work but change the value of the card permanently. created a test card that can be used to test the upgrade room with buying and upgrading the values.